requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
loguru==0.7.2
cohere

# LLM / RAG related
langchain==0.1.20
sentence-transformers==2.5.1
pinecone==6.0

# Optional extras
openai==1.23.2   # If fallback or embeddings are from OpenAI
tqdm==4.66.2